feature: add department controller
