+++ backend/BackendMIIT/src/main/java/com/example/BackendMIIT/controller/DepartmentController.java
@@ -1,0 +1,10 @@
+package com.example.BackendMIIT.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(path = "/api/miit/departments")
+public class DepartmentController {
+
+}
