+++ backend/BackendMIIT/src/main/java/com/example/BackendMIIT/model/domain/Department.java
@@ -1,0 +1,14 @@
+package com.example.BackendMIIT.model.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.Data;
+
+@Entity
+@Table(name = "departments")
+@Data
+public class Department extends BaseEntity {
+
+    private String name;
+    private String teacher;
+}
