+++ backend/BackendMIIT/src/main/java/com/example/BackendMIIT/model/domain/Exam.java
@@ -1,0 +1,19 @@
+package com.example.BackendMIIT.model.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Data;
+
+@Entity
+@Table(name = "exams")
+@Data
+public class Exam extends BaseEntity {
+
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "direction_id")
+    private Direction direction;
+}
