+++ backend/BackendMIIT/src/main/java/com/example/BackendMIIT/model/dto/DirectionDto.java
@@ -1,14 +1,14 @@
-package com.example.BackendMIIT.models.dto;
+package com.example.BackendMIIT.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
-public class DepartmentDto {
+public class DirectionDto {
 
     @JsonProperty("name")
     private String name;
 
-    @JsonProperty("teacher")
-    private String teacher;
+    @JsonProperty("pass_point")
+    private int passPoint;
 }
