--- backend/BackendMIIT/src/main/java/com/example/BackendMIIT/models/domain/BaseEntity.java
@@ -1,16 +1,0 @@
-package com.example.BackendMIIT.models.domain;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
-}
