--- backend/BackendMIIT/src/main/java/com/example/BackendMIIT/models/domain/Department.java
@@ -1,14 +1,0 @@
-package com.example.BackendMIIT.models.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.Data;
-
-@Entity
-@Table(name = "departments")
-@Data
-public class Department extends BaseEntity {
-
-    private String name;
-    private String teacher;
-}
